salmasimpleshell

+++ handl_builtin.c
@@ -43,7 +43,7 @@ void print_env(char **command, int **status)
     int i;
     for (i = 0; environ[i]; i++)
     {
-        write(STDOUT_FILENO, environ[i], _strlen(environ[i]));
+        write(STDOUT_FILENO, environ[i], (size_t)_strlen(environ[i]));
         write(STDOUT_FILENO, "\n", 1);
     }
     freearystring(command);
